PDFBOX-4497: remove cast to keep it a float, adjust type
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1856432 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/rendering/PageDrawer.java
@@ -633,9 +633,9 @@ public class PageDrawer extends PDFGraphicsStreamEngine
         }
 
         PDLineDashPattern dashPattern = state.getLineDashPattern();
-        int phaseStart = dashPattern.getPhase();
+        float phaseStart = dashPattern.getPhase();
         float[] dashArray = getDashArray(dashPattern);
-        phaseStart = (int) transformWidth(phaseStart);
+        phaseStart = transformWidth(phaseStart);
 
         // empty dash array is illegal
         // avoid also infinite and NaN values (PDFBOX-3360)
